refactor(client-telemetry): split primitives.rs under the 700-line cap
primitives.rs was ~781 lines, over the 700-line hard cap in CLAUDE.md.
Move the `#[cfg(all(test, target_arch = "x86"))] mod tests` block into a
sibling `tests.rs` and convert primitives.rs -> primitives/mod.rs so
`mod primitives;` in hooks.rs keeps resolving. mod.rs is now 590 lines.

Pure move: the primitives impl and the test logic are byte-identical
(only re-indented one level by rustfmt now that the tests are top-level
in their own file). Also drops the stray issue-number reference from the
module doc comment per the no-#NNN-in-tracked-source convention.
